refactor: move debugger files into a dedicated pkg
