+++ internal/debugger/agent_pod_view_test.go
@@ -1,4 +1,4 @@
-package main
+package debugger
 
 import (
 	"testing"
