+++ internal/debugger/cluster_pod_view.go
@@ -1,4 +1,4 @@
-package main
+package debugger
 
 import (
 	"fmt"
