+++ internal/debugger/validate_test.go
@@ -1,4 +1,4 @@
-package main
+package debugger
 
 import (
 	"bytes"
